+++ examples/src/main/java/com/mapbox/navigation/examples/core/MapboxManeuverActivity.kt
@@ -132,9 +132,6 @@ class MapboxManeuverActivity : AppCompatActivity(), OnMapLongClickListener {
     }
 
     private val locationObserver = object : LocationObserver {
-
-        override fun onNewRawLocation(rawLocation: Location) {}
-
         override fun onNewLocationMatcherResult(locationMatcherResult: LocationMatcherResult) {
             navigationLocationProvider.changePosition(
                 locationMatcherResult.enhancedLocation,